Add more built-in colors
This adds all named HTML colors to be supported by MarkWhen tags. In the event of a collision, the MarkWhen selected value will be used instead of the HTML value. The previously declared `COLORS` array has been renamed to `MARKWHEN_COLORS` and is used in the other same places, mainly to cycle through fallabck colors if a user-specified color is wrong.

This also adds a new tag color syntax: `html(<name>)`. This can be used to force selection of the HTML-variant of the color name instead of the MarkWhen-variant.